3 Use proper init process
Added use of supervisord to control processes.
supervisord is mentioned on https://docs.docker.com/engine/admin/multi-service_container/ as a moderately heavy-weight approach, a huge advantage (over even more lightweight process managers) is that it can easily be installed via apt-get.
The supervisord config file is created inside the Dockerfile to have all information in one place, this could be changed for the sake of better organisation.